+++ spinnman/extended/__init__.py
@@ -1,0 +1,1 @@
+from .de_alloc_sdram_process import DeAllocSDRAMProcess
